+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:animated_splash_screen/animated_splash_screen.dart';
 import 'package:chalynyx_todo_app/routes.dart';
-import 'package:chalynyx_todo_app/splash.dart';
+import 'package:chalynyx_todo_app/screens/splash.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
